requirements.txt
cachetools==3.0.0
certifi==2018.8.24
chardet==3.0.4
defusedxml==0.5.0
Django==2.1.3
django-materializecss-form==1.1.10
entrypoints==0.2.3
google-api-python-client==1.7.5
google-auth==1.6.1
google-auth-httplib2==0.0.3
gunicorn==19.9.0
httplib2==0.12.0
idna==2.7
keyring==15.0.0
mysqlclient==1.3.13
oauthlib==2.1.0
pip-autoremove==0.9.1
pkginfo==1.4.2
pyasn1==0.4.4
pyasn1-modules==0.2.2
PyJWT==1.6.4
python3-openid==3.1.0
pytz==2018.5
requests==2.19.1
requests-oauthlib==1.0.0
requests-toolbelt==0.8.0
rsa==4.0
six==1.11.0
social-auth-app-django==3.1.0
social-auth-core==2.0.0
tqdm==4.25.0
twine==1.11.0
uritemplate==3.0.0
urllib3==1.23
